refactor: Refatorado localização de classes
